feat: refactor feedback and footer to components
- removed footer and feedback from App.tsx to individual components
- added all text sections to translation.json for clean structure
