+++ src/i18n/translations/translations.json
@@ -32,6 +32,10 @@
     "intro.p2": "Du hast noch keinen Termin?",
     "intro.p2.link": "Termin buchen.",
 
+    "feedback.question": "Wie gefällt Dir die Anwendung?",
+    "feedback.link": "https://citylabberlin.typeform.com/to/kCdnCgvC?product_id=checklist",
+    "feedback.link.label": "Beantworte uns gerne einige Fragen.",
+
     "first-registration.q1": "Meldest Du Dich zum ersten Mal in Berlin an?",
     "first-registration.q1.tooltip": "Wenn ja, müssen eventuell Dokumente wie Geburts- oder Heiratsurkunde vorgezeigt werden.",
     "first-registration.q2": "Bist Du verheiratet?",
@@ -104,7 +108,14 @@
     "nonEuIdOrPassportOrReplacement": "Reisepass oder Passersatzpapiere",
     "nonEuIdOrPassportOrReplacement.tooltip": "Dieses Dokument ist notwendig, um Dich auszuweisen.",
     "supplement": "Beiblatt zur Anmeldung/Hauptwohnungserklärung",
-    "supplement.tooltip": "Dieses Dokument ist notwendig, weil Du weiterhin in einer weiteren Wohnung gemeldet bleiben möchtest."
+    "supplement.tooltip": "Dieses Dokument ist notwendig, weil Du weiterhin in einer weiteren Wohnung gemeldet bleiben möchtest.",
+
+    "logo.t1": "Durchgeführt von",
+    "logo.t2": "Ein Projekt der",
+    "logo.t3": "Mit Unterstützung von",
+
+    "dataProtection": "Datenschutz",
+    "imprint": "Impressum"
   },
   "en-DE": {
     "de-DE.mobile": "DE",
@@ -138,6 +149,10 @@
     "intro.p1": "This application is a prototype from CityLAB Berlin and is currently being used for a test. It should help you to prepare well for your application appointment. Answer a few questions and we will provide you with an overview of the documents you need to bring with you to the appointment. Let's go!",
     "intro.p2": "Don't have an appointment yet?",
     "intro.p2.link": " Book an appointment.",
+
+    "feedback.question": "How do you like the application?",
+    "feedback.link": "https://citylabberlin.typeform.com/to/kCdnCgvC?product_id=checklist",
+    "feedback.link.label": "Feel free to give us some feedback",
 
     "first-registration.q1": "Are you registering in Berlin for the first time?",
     "first-registration.q1.tooltip": "If yes, documents such as a birth or marriage certificate may need to be presented.",
@@ -211,6 +226,13 @@
     "nonEuIdOrPassportOrReplacement": "Passport or Passport Replacement Documents",
     "nonEuIdOrPassportOrReplacement.tooltip": "This document is necessary to identify yourself.",
     "supplement": "Supplement to Registration/Main Residence Declaration",
-    "supplement.tooltip": "This document is necessary because you want to remain registered in another apartment."
+    "supplement.tooltip": "This document is necessary because you want to remain registered in another apartment.",
+
+    "logo.t1": "Executed by",
+    "logo.t2": "A project by",
+    "logo.t3": "Supported by",
+
+    "dataProtection": "Data protection",
+    "imprint": "Imprint"
   }
 }
